requirements.txt
# requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pandas>=1.2.4
Pillow>=8.2.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1